backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyJWT>=2.8.0
anthropic>=0.31.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
openai>=1.35.0
pinecone>=3.2.0
python-multipart>=0.0.9
slowapi>=0.1.9
